requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
black==24.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
frozenlist==1.4.1
furl==2.1.3
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
multidict==6.0.5
multidimensional-urlencode==0.0.4
mypy-extensions==1.0.0
nose==1.3.7
orderedmultidict==1.0.1
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pook==1.4.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
referencing==0.33.0
requests==2.31.0
retailcrm==5.1.2
rpds-py==0.17.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
xmltodict==0.13.0
yarl==1.9.4
